Move calc_pgh to separate files
