+++ projection_matrix_gpu.py
@@ -1,4 +1,5 @@
 import cupy as cp
+import cupyx.scipy.special
 from numba import cuda
 
 def native_endian(data):
@@ -105,3 +106,73 @@ def evalcoeffs(wavelengths, psfdata):
         p[key] = psfdata.meta[key]
 
     return p
+
+def calc_pgh(ispec, wavelengths, psfparams):
+    '''
+    Calculate the pixelated Gauss Hermite for all wavelengths of a single spectrum
+
+    ispec : integer spectrum number
+    wavelengths : array of wavelengths to evaluate
+    psfparams : dictionary of PSF parameters returned by evalcoeffs
+
+    returns pGHx, pGHy
+
+    where pGHx[ghdeg+1, nwave, nbinsx] contains the pixel-integrated Gauss-Hermite polynomial
+    for all degrees at all wavelengths across nbinsx bins spaning the PSF spot, and similarly
+    for pGHy.  The core PSF will then be evaluated as
+
+    PSFcore = sum_ij c_ij outer(pGHy[j], pGHx[i])
+    '''
+
+    #- shorthand
+    p = psfparams
+
+    #- spot size (ny,nx)
+    nx = p['HSIZEX']
+    ny = p['HSIZEY']
+    nwave = len(wavelengths)
+    p['X'], p['Y'], p['GHSIGX'], p['GHSIGY'] = \
+    cp.array(p['X']), cp.array(p['Y']), cp.array(p['GHSIGX']), cp.array(p['GHSIGY'])
+    xedges = cp.repeat(cp.arange(nx+1) - nx//2, nwave).reshape(nx+1, nwave)
+    yedges = cp.repeat(cp.arange(ny+1) - ny//2, nwave).reshape(ny+1, nwave)
+
+    #- Shift to be relative to the PSF center at 0 and normalize
+    #- by the PSF sigma (GHSIGX, GHSIGY)
+    #- xedges[nx+1, nwave]
+    #- yedges[ny+1, nwave]
+    xedges = (xedges - p['X'][ispec]%1)/p['GHSIGX'][ispec]
+    yedges = (yedges - p['Y'][ispec]%1)/p['GHSIGY'][ispec]
+
+    #- Degree of the Gauss-Hermite polynomials
+    ghdegx = p['GHDEGX']
+    ghdegy = p['GHDEGY']
+
+    #- Evaluate the Hermite polynomials at the pixel edges
+    #- HVx[ghdegx+1, nwave, nx+1]
+    #- HVy[ghdegy+1, nwave, ny+1]
+    HVx = hermevander_wrapper(xedges, ghdegx).T
+    HVy = hermevander_wrapper(yedges, ghdegy).T
+
+    #- Evaluate the Gaussians at the pixel edges
+    #- Gx[nwave, nx+1]
+    #- Gy[nwave, ny+1]
+    Gx = cp.exp(-0.5*xedges**2).T / cp.sqrt(2. * cp.pi)
+    Gy = cp.exp(-0.5*yedges**2).T / cp.sqrt(2. * cp.pi)
+
+    #- Combine into Gauss*Hermite
+    GHx = HVx * Gx
+    GHy = HVy * Gy
+
+    #- Integrate over the pixels using the relationship
+    #  Integral{ H_k(x) exp(-0.5 x^2) dx} = -H_{k-1}(x) exp(-0.5 x^2) + const
+
+    #- pGHx[ghdegx+1, nwave, nx]
+    #- pGHy[ghdegy+1, nwave, ny]
+    pGHx = cp.zeros((ghdegx+1, nwave, nx))
+    pGHy = cp.zeros((ghdegy+1, nwave, ny))
+    pGHx[0] = 0.5 * cp.diff(cupyx.scipy.special.erf(xedges/cp.sqrt(2.)).T)
+    pGHy[0] = 0.5 * cp.diff(cupyx.scipy.special.erf(yedges/cp.sqrt(2.)).T)
+    pGHx[1:] = GHx[:ghdegx,:,0:nx] - GHx[:ghdegx,:,1:nx+1]
+    pGHy[1:] = GHy[:ghdegy,:,0:ny] - GHy[:ghdegy,:,1:ny+1]
+    
+    return pGHx, pGHy
